pyproject.toml
[tool.poetry]
name = "foundation-python"
version = "1.0.2"
description = ""
authors = ["bai ying <cn.baiying@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
bullmq = "^2.9.3"
redis = "^5.0.8"
python-dotenv = "^1.0.1"
shortuuid = "^1.0.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
